+++ packages/block-library/src/more/index.js
@@ -9,16 +9,12 @@ import { compact } from 'lodash';
 import { __, _x } from '@wordpress/i18n';
 import { RawHTML } from '@wordpress/element';
 import { createBlock } from '@wordpress/blocks';
-import {
-	Path,
-	G,
-	SVG,
-} from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
 import edit from './edit';
+import icon from './icon';
 
 export const name = 'core/more';
 
@@ -27,7 +23,7 @@ export const settings = {
 
 	description: __( 'Content before this block will be shown in the excerpt on your archives page.' ),
 
-	icon: <SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><Path fill="none" d="M0 0h24v24H0V0z" /><G><Path d="M2 9v2h19V9H2zm0 6h5v-2H2v2zm7 0h5v-2H9v2zm7 0h5v-2h-5v2z" /></G></SVG>,
+	icon,
 
 	category: 'layout',