+++ packages/block-library/src/quote/contants.js
@@ -1,0 +1,2 @@
+export const ATTRIBUTE_QUOTE = 'value';
+export const ATTRIBUTE_CITATION = 'citation';